cover timezone and automotive
